admin view updates
