+++ project/apps/schedules/models/__init__.py
@@ -1,3 +1,4 @@
-from .order_schedule_models import GravelDeliverySchedule, PumpSchedule
+from .concrete_pump_schedule import PumpSchedule
+from .gravel_delivery_schedule import GravelDeliverySchedule
 from .schedule_delay_models import InclimateWeather, InclimateWeatherNote
 from .schedule_notes_models import GravelDeliveryScheduleNote, PumpScheduleNote
